+++ changelog.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - structurizr-core: `ViewSet.isEmpty()` was missing a check for image views.
+- structurizr-dsl: Fixes https://github.com/structurizr/java/issues/252 (DSL parser does not seem to handle curly brackets balance).
 
 ## 2.0.0 (22nd February 2024)